Ensure routes that do not contain a "?" mark show up in -report
The route_regexp expression was expecting routes to be terminated by a "?", this missed routes that passed no GET parameters. Improved the regexp to look for ? or "

Used https://regexr.com/ to test the expressions and a msg("no matching route for new report! " + line); in the else of the "if route_match" to figure out when it was not matching.