Update request and message regexp to handle log lines with request identifiers that end in RA/RS/AT
These suffixes were added to the request identifers to communicate if logs were coming from other threads and for sync/async mounts. I noticed that if a match was found on one of these lines, it did not print the rest of the lines from the request when using -full.

If you want to match a line such as:
[2021-07-13 17:57:28 UTC] P8568R135000RA ERROR   RvSphere: vSphere operation failed: Invalid configuration for device '0'.

This command will now print all lines "P8568R135000" lines even if they have suffixes
> avmlog -full -find='Invalid configuration for device' file.log
[2021-07-13 17:57:20 UTC] P8568R135000 ....
[2021-07-13 17:57:30 UTC] P8568R135000RA ....
[2021-07-13 17:57:40 UTC] P8568R135000AT ....
[2021-07-13 17:57:50 UTC] P8568R135000 ....